requirements.txt
brotlipy==0.7.0
cffi==1.11.2
confusable-homoglyphs==2.0.2
Django==2.0
django-bootstrap3==9.1.0
django-countries==4.4
django-crispy-forms==1.7.0
django-registration==2.3
django-stdimage==3.0.0
djangorestframework==3.7.3
olefile==0.44
Pillow==4.3.0
progressbar2==3.34.3
psycopg2==2.7.3.2
pycparser==2.18
python-utils==2.2.0
pytz==2017.3
six==1.11.0
whitenoise==3.3.1
